Ensure proper handling of UTF-8 strings in HMAC
@wslack hit this bug when updates to 18F/partnership-playbook weren't
propagating to https://pages.18f.gov/partnership-playbook/. I found out that
the HMAC signature for the webhook didn't match what was sent by GitHub.

After verifying the exact same inputs and computing the GitHub version of the
signature locally in a Ruby script, I eventually noticed that adding the
'utf8' encoding specifier to Hmac.update() solved the problem. I then was able
to identify that the trouble-making part of the payload was an instance of the
string `it’s` where the apostrophe is a UTF-8 smart quote character.